backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
requests>=2.31.0
pandas>=2.0.3
numpy>=1.24.3
yfinance>=0.2.18
akshare>=1.11.0
tushare>=1.3.0
openai>=1.12.0
python-dotenv>=1.0.0
pytz
ta>=0.10.2
peewee>=3.17.0
schedule>=1.2.0
pywencai>=0.7.0
jieba>=0.42.1
pymongo>=4.6.1
redis>=5.0.1
